no-cpld: new overlay
Once we have a full featured CPLD driver, the CPLD_INTERRUPT# gpio will
be claimed by the driver and will be unavailable to the userspace. Eg.
altera-stapl won't work then anymore. Be prepared for that scenario and
provide an overlay which can disable the CPLD support on the board. This
only works with the base device tree, you should not load any other
carrier device tree overlay together with this overlay.

Signed-off-by: Michael Walle <michael.walle@kontron.com>